feat(devenv): updated folders and added sln
- updated folder structure
- introduced a sln file
- added solution explorer VS code plugin to recommended plugins
- updated settings and docker related files
- tested the app again
- updated the readme
