+++ .vscode/extensions.json
@@ -13,7 +13,8 @@
 		"ms-vscode-remote.remote-containers",
 		"patcx.vscode-nuget-gallery",
 		"Gruntfuggly.todo-tree",
-		"wayou.vscode-todo-highlight"
+		"wayou.vscode-todo-highlight",
+		"fernandoescolar.vscode-solution-explorer"
 	],
 	// List of extensions recommended by VS Code that should not be recommended for users of this workspace.
 	"unwantedRecommendations": [
