+++ Backend.sln
@@ -1,0 +1,48 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.6.30114.105
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Todo.Backend.Tests", "Todo.Backend.Tests\Todo.Backend.Tests.csproj", "{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Todo.Backend", "Todo.Backend\Todo.Backend.csproj", "{136328C0-6BA6-4DAC-9029-1A0A9369EB27}"
+EndProject
+Global
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|Any CPU = Release|Any CPU
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|x64.Build.0 = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Debug|x86.Build.0 = Debug|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|Any CPU.Build.0 = Release|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|x64.ActiveCfg = Release|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|x64.Build.0 = Release|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|x86.ActiveCfg = Release|Any CPU
+		{015C2EAD-46FF-44C2-A7F0-02C518EA8A62}.Release|x86.Build.0 = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|x64.Build.0 = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Debug|x86.Build.0 = Debug|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|Any CPU.Build.0 = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|x64.ActiveCfg = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|x64.Build.0 = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|x86.ActiveCfg = Release|Any CPU
+		{136328C0-6BA6-4DAC-9029-1A0A9369EB27}.Release|x86.Build.0 = Release|Any CPU
+	EndGlobalSection
+EndGlobal
