+++ Todo.Backend.Tests/CommandTests.cs
@@ -1,0 +1,9 @@
+using System;
+using Xunit;
+
+namespace Backend.Tests
+{
+    public class CommandTests
+    {
+    }
+}
