+++ Todo.Backend/Commands/DeleteTodoItemCommand.cs
