+++ Todo.Backend/Commands/DeleteTodoItemCommandHandler.cs
