+++ Todo.Backend/Dockerfile
@@ -5,16 +5,16 @@ EXPOSE 5000
 
 FROM mcr.microsoft.com/dotnet/sdk:5.0 AS build
 WORKDIR /src
-COPY ["backend.csproj", "./"]
-RUN dotnet restore "backend.csproj"
+COPY ["Todo.Backend.csproj", "./"]
+RUN dotnet restore "Todo.Backend.csproj"
 COPY . .
 WORKDIR "/src/."
-RUN dotnet build "backend.csproj" -c Release -o /app/build
+RUN dotnet build "Todo.Backend.csproj" -c Release -o /app/build
 
 FROM build AS publish
-RUN dotnet publish "backend.csproj" -c Release -o /app/publish
+RUN dotnet publish "Todo.Backend.csproj" -c Release -o /app/publish
 
 FROM base AS final
 WORKDIR /app
 COPY --from=publish /app/publish .
-ENTRYPOINT ["dotnet", "backend.dll"]
+ENTRYPOINT ["dotnet", "Todo.Backend.dll"]
