+++ Todo.Backend/Persistence/Entities/TodoItem.cs
